Made changes so that ajax call would be made as a "GET"
Made a change so that ajax call is a "GET" rather than a "POST". Also changed the "called =false" in the ajax call. Please double check that these are the only changes I made before pulling (I worked on this in several different chunks and may have made some other changes that I do not remember)