Added vendor_id and vendor_name
Added option of how you receive vendor_id and vendor_name. Can be achieved by taking the comments off of the 2 echo statements and commenting the single echo statement. The single echo gives and output such that "vendor_id Vendor name" and the 2 echo statements give an output such that "vendor_id" "vendor_name". choose whichever will be best for you when parsing and saving the vendor_id for use with the selected vendor name in the form.