Gave name to submit button.
Gave a name to the submit button. That's all it was. An unnamed god d**n submit button. We can now add orders to the database with a relationship to the vendor table. Will add functionality to the login page later on so that the username and id of the logged in user is also submitted with it.